comments were fixed
